Add merge option instead of sync.
Merge will use patch instead of replace to be able to merge configMaps or secrets from different namespaces with different keys to one "global" configMap. This was the initial behavior for synator.

Merge uses a new label 'synator/merge' and does not watch for delete events. If some keys must be deleted, it should be done manually.

This can be used to generate a global configMap with different keys from different namespaces. That configMap could be mounted to a directory and different namespaces can add different files to the directory for example.